requirements.txt
django
djangorestframework
Pillow
django-mptt
django-reset-migrations
mysqlclient
django-cors-headers